requirements.txt
autogen-agentchat
autogen-ext[openai,azure,mcp]
openai
python-dotenv
httpx
aiohttp
mcp[cli]
uvicorn
fastapi